Workaround an issue in some browsers with the rich text editor
In some browsers, the text cursor in the Markdown editor widget can jump while typing in certain circumstances, making it more or less impossible to use. This was reported in https://github.com/netlify/netlify-cms/issues/5092 which also documents a workaround. This commit adds the workaround to the admin dashboard so we can benefit from it by default.